requirements.txt
django==1.10
django-phonenumber-field
psycopg>=2.6